+++ lib/pages/3pSettings.dart
@@ -5,6 +5,7 @@ import 'package:libacflutter/Constants.dart';
 import 'package:markdown_widget/widget/markdown_block.dart';
 import 'package:switchboard/dart/globalHelpers.dart';
 import 'package:switchboard/dart/octocon_format.dart';
+import 'package:switchboard/dart/ourcana_format.dart';
 import 'package:switchboard/dart/pluralkit_format.dart';
 import 'package:switchboard/dart/storage.dart';
 import 'package:switchboard/file_picker.dart';
@@ -33,18 +34,31 @@ class _octocon extends State<OctoconImport> {
     bool fail = false;
     try {
       octoconData = OctoconData.fromJson(octoData);
-    } catch (E) {
-      // Maybe it is in PluralKit format?
+    } catch (E, stack) {
+      print(E);
+      print(stack);
       fail = true;
       try {
         PluralKitData pkData = await PluralKit.decode(
           typeCorrectJsonDecode(octoData),
         );
 
-        fail = false;
         octoconData = PluralKit.convertToOctocon(pkData);
-      } catch (E) {
+        fail = false;
+      } catch (E, stack) {
+        print(E);
+        print(stack);
         // Maybe try Ourcana next?
+        try {
+          OurcanaData ocData = Ourcana.decode(typeCorrectJsonDecode(octoData));
+
+          octoconData = Ourcana.convertToOctocon(ocData);
+          fail = false;
+        } catch (E, stack) {
+          print(E);
+          print(stack);
+          // No other supported Codecs
+        }
       }
     }
 
@@ -52,10 +66,11 @@ class _octocon extends State<OctoconImport> {
       ScaffoldMessenger.of(context).showSnackBar(
         SnackBar(
           content: Text(
-            "Fatal Error: The provided file could not be decoded using any available codec. We attempted Octocon, and PluralKit. No valid format succeeded. If you believe this to be an error, please contact the development team. You may be asked to provide the export json in order for us to better assist you.",
+            "Fatal Error: The provided file could not be decoded using any available codec. We attempted Octocon, PluralKit, and Ourcana. No valid format succeeded. If you believe this to be an error, please contact the development team. You may be asked to provide the export json in order for us to better assist you.",
           ),
         ),
       );
+      return;
     }
 
     await Navigator.pushNamed(
@@ -87,7 +102,7 @@ class _octocon extends State<OctoconImport> {
             children: [
               MarkdownBlock(
                 data:
-                    "# Welcome\n\nThis is the import system. You will be able to import from a 3rd party here.\n\n## Octocon\n\nOctocon full exports are completely supported, as are PluralKit exports from Octocon. Please be aware, only a full export will have your fronting history.\n\n## PluralKit\n\nPluralKit support is experimental at best. We partially support this platform, but the PK JSON file we were provided by a friend did not contain any fronting history, so PK only is able to import your system, no history.\n\n# Option 1: Complete Wipe\n\nThis option will completely wipe your Switchboard account and replace its contents with that of your imported file. This is for you if you wanted to try out our app before importing everything.\n\n# Option 2: As Is\n\nThis option is for you if you want to combine profile systems. It will merge the data from an export into Switchboard as seamlessly as we possibly can. Again, this option is non-destructive and will **not** erase anything.",
+                    "# Welcome\n\nThis is the import system. You will be able to import from a 3rd party here.\n\n## Octocon\n\nOctocon full exports are completely supported, as are PluralKit exports from Octocon. Please be aware, only a full export will have your fronting history.\n\n## PluralKit\n\nPluralKit support is experimental at best. We partially support this platform, but the PK JSON file we were provided by a friend did not contain any fronting history, so PK only is able to import your system, no history.\n\n## Ourcana\n\nOurcana support is experimental at best. We partially support this platform, however the JSON file we had available to examine did not contain every possible information type. So, some things like custom fields may not work.\n\n# Option 1: Complete Wipe\n\nThis option will completely wipe your Switchboard account and replace its contents with that of your imported file. This is for you if you wanted to try out our app before importing everything.\n\n# Option 2: As Is\n\nThis option is for you if you want to combine profile systems. It will merge the data from an export into Switchboard as seamlessly as we possibly can. Again, this option is non-destructive and will **not** erase anything.",
               ),
               Divider(),
               ListTile(