+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## ADD:
 
 - [x] PluralKit exports are now partially supported.
+- [x] Ourcana exports are now partially supported.
 
 # 0.3.2+0711261047 HOTFIX